refactor(data): FullIMURaw SensorData and Entity now stores all values as floats
The conversion is a hardcoded division with 16384 for acceleration and
131 for gyroscope which ONLY makes sense for MPU6050 sensor. This will
not work if the sensor is configured with other full-scale range values
or if you are using a different sensor

+++ app/src/main/java/com/example/esp32_mpu6050_mobile_data_collection/data/database/entity/FullIMURawEntity.kt
@@ -22,13 +22,13 @@ import androidx.room.PrimaryKey
 data class FullIMURawEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val sessionId: Long, // Foreign Key
-    val ax: Short,
-    val ay: Short,
-    val az: Short,
+    val ax: Float,
+    val ay: Float,
+    val az: Float,
 
-    val gx: Short,
-    val gy: Short,
-    val gz: Short,
+    val gx: Float,
+    val gy: Float,
+    val gz: Float,
 
     val qx: Float,
     val qy: Float,
